+++ src/environment.h
@@ -15,14 +15,14 @@ public:
     std::tuple<int, int, bool> Step(const std::string& action);
     std::vector<std::string> actions;
     int num_states;
-    int num_actions;
+    size_t num_actions;
 private:
     int CalculateReward();
     bool CheckTermination();
     std::vector<std::string> states;
     int current_state = std::distance(states.begin(), std::find(states.begin(), states.end(), "neutral"));
-    int days_lived = 0;
-    int days_without_eating = 0;
-    int max_days = 50;
-    int max_days_without_eating = 43;
+    size_t days_lived = 0;
+    size_t days_without_eating = 0;
+    size_t max_days = 50;
+    size_t max_days_without_eating = 43;
 };
